Finished API, added timing checks and roundtrip checks.
Added a run method which takes and receives no arguments for scripts.

Added error reporting to the eval loop by using try/catch and lasterr()

Added a check for the raw speed of the library, comparing tic/toc to
our overheads - which seem to be less than 50% unless we are passing
large datasets around repeatedly.

Added a check for round trip preserving of variable value and type.
Did not yet incorporate cell arrays and strings.