Moved to .mat files for Python 3 support to pick up types unsupported by HDF.
HDF files written in Python 3 could not be read in Octave 3.2.4 for Windows,
most likely due to the fact that a 1.6 HDF library is used.

Also, cell arrays could not be recreated using an HDF file, whereas
scipy.savemat could handle them fine.

Scipy.savemat was chosen for all file writes, keeping HDF for file reads.
Larger arrays take about twice as long to save, but the increase in functionality
was considered an appropriate trade.