FIX: retrhrow() mat-side errors not to loose context
As hinted in in my recent comment #115.
i'm losing the location context of errors from the matlab-side.

I traced the error to the matlab statement used to re-raise any errors caught.
The proper way is to `rethrow()` errors, not to raise them as new errors.
Did it work in the older versions (from mine, Octave-5.2.0 & oct2py 5.0.4)?